pyproject.toml
[project]
name = "agent-recall"
version = "0.1.0"
description = "Persistent memory for AI coding agents"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "Your Name", email = "you@example.com" },
]
keywords = ["ai", "agents", "memory", "llm", "coding"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]

dependencies = [
  "typer>=0.9.0",
  "pydantic>=2.0",
  "structlog>=24.0",
  "rich>=13.0",
  "pyyaml>=6.0",
  "anthropic>=0.40",
  "openai>=1.0",
]

[project.optional-dependencies]
google = ["google-generativeai>=0.5"]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "ruff>=0.8",
  "ty>=0.0.1a13",
]

[project.scripts]
agent-recall = "agent_recall.cli.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_recall"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
